+++ with.go
@@ -26,7 +26,6 @@ func (d *Driver) with(ctx context.Context, opts ...Option) (*Driver, uint64, err
 					delete(d.children, id)
 				}),
 				withConnPool(d.pool),
-				WithSharedBalancer(d),
 			),
 			opts...,
 		)...,