+++ driver.go
@@ -415,10 +415,11 @@ func (d *Driver) connect(ctx context.Context) (err error) {
 	if d.pool == nil {
 		d.pool = conn.NewPool(ctx, d.config)
 	}
-
-	d.balancer, err = balancer.New(ctx, d.config, d.pool, d.discoveryOptions...)
-	if err != nil {
-		return xerrors.WithStackTrace(err)
+	if d.balancer == nil {
+		d.balancer, err = balancer.New(ctx, d.config, d.pool, d.discoveryOptions...)
+		if err != nil {
+			return xerrors.WithStackTrace(err)
+		}
 	}
 
 	d.table = xsync.OnceValue(func() (*internalTable.Client, error) {