+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Use common balancer for child drivers
+
 ## v3.89.0
 * Fixed send optional arguments to the server with `ydb.ParamsBuilder`