+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Use common balancer for child drivers
+* Added option `ydb.WithSharedBalancer(*Driver)` for child drivers
 
 ## v3.89.0
 * Fixed send optional arguments to the server with `ydb.ParamsBuilder`